fix: use UInt32 for duration_ms and improve ORDER BY in execution engine tables
- Change duration_ms from UInt64 to UInt32 (consistent with older tables)
- Update execution_engine_new_payload ORDER BY to use block_number first

+++ deploy/migrations/clickhouse/091_execution_engine_get_blobs.up.sql
@@ -5,7 +5,7 @@ CREATE TABLE execution_engine_get_blobs_local ON CLUSTER '{cluster}' (
   requested_date_time DateTime64(3) COMMENT 'Timestamp when the engine_getBlobs call was received' Codec(DoubleDelta, ZSTD(1)),
 
   -- Timing
-  duration_ms UInt64 COMMENT 'How long the engine_getBlobs call took in milliseconds' Codec(ZSTD(1)),
+  duration_ms UInt32 COMMENT 'How long the engine_getBlobs call took in milliseconds' Codec(ZSTD(1)),
 
   -- Source
   source LowCardinality(String) COMMENT 'Source of the event (SNOOPER, EXECUTION_CLIENT)',
